Patched catastrophic leak of Tetriminos
After the grand Purge of malloc, our people rejoiced. But not for long, for a lone leak spy had sprung a leak of tetriminos in the great wall of Fillit.
Due to minimal avaiable building space in the great wall of Fillit, we had to improvise. But rejoice, for our great engineers have patched said leak within the present limitations! They cleverly relocated the check for amount of Tetrominos inside the grat wall of Fillit from the outer wall, to the inner corridors.

Liewe Gutter
 -Monday 1st, April 2019