Finalization of optimisation
After the loss of the battle in the eval pool on April 1st, we were able to re-motivate our troops by assigning them with new optimisation tasks. This experiment was a succes, resulting in numerous new optimisations, improving agility of our infantry, and solidifying modifiers. This will greatly improve our chances in tonights battle!

-Liewe Gutter, Wednesday April 3rd