further improvement of Makefiles
After the succes of the last experiment, and with todays battle in sight, the troops had excess energy. They begged us to provided them with an additional challenge, which we provided. They managed to make our management systems even MORE efficient, providing much better tracking of time needed for management tasks. A great victory for our army, as this is sure to make our adversaries die of jealousy!

-Liewe Gutter, Wednesday April 3rd